formating

+++ clean_slate.c
@@ -32,7 +32,7 @@
         }\
     } while(0);
 
-enum FailStatus {
+enum FinishStatus {
     FAILURE_EXIT    = 0,     
     FALIURE_RETRY   = 1,     
     SUCCESS_EXIT    = 2     
@@ -40,58 +40,81 @@ enum FailStatus {
 
 typedef enum WipeAlgs {
     USE_NONE        = 0,
-    USE_ZEROS       = 1, 
-    USE_ONES        = 2,
-    USE_PSEUDO      = 3,
-    USE_GOST        = 4,
-    USE_AIRFORCE    = 5,
-    USE_ARMY        = 6,
-    USE_HMG         = 7,
-    USE_DOD         = 8,
-    USE_PFITZNER    = 9,
-    USE_GUTMANN     = 10,
-    USE_SOURCE      = 11
+    USE_ZEROS       = 1,  //-z    
+    USE_ONES        = 2,  //-o
+    USE_PSEUDO      = 3,  //-p
+    USE_GOST        = 4,  //-r
+    USE_AIRFORCE    = 5,  //-c
+    USE_ARMY        = 6,  //-a
+    USE_HMG         = 7,  //-b
+    USE_DOD         = 8,  //-d
+    USE_PFITZNER    = 9,  //-n
+    USE_GUTMANN     = 10, //-g
+    USE_SOURCE      = 11  //-s
 } WipeAlgs;
 
 static WipeAlgs wipe_with;
 
 
 typedef struct Options {
-    bool displ_info;
-    bool delete_after;
-    bool force_open;
-    int fd_src;
-    int first_last;
-    int file_size_bytes;
-    int firs_fp;
-    int last_fp;
+    bool delete_after; //-e
+    bool displ_info;   //-i
+    bool force_open;   //-f
+    int first;    //-t ( 0 if flag is not set)
 } Options;
 
 static Options options;
 
+int first_fp;   
+int last_fp;
+int size;
 
 typedef struct w_info {
     bool wipe_failed;
-    char **str_errors_buf;
-    int buf_length;
-    int prec;
+    char **errors;
+    int n_errors;
+    int completed_prec;
     time_t start_tm;
-    time_t end_time;
+    time_t end_tm;
 } w_info;
 
 
-struct option const long_opts[] = {
+const struct option long_opts[] = {
+    {"source", required_argument, NULL, 's'}, 
     {"help"  , no_argument      , NULL, 'h'},
-    {"source", required_argument, NULL, 's'},
     {"first" , required_argument, NULL, 't'}
 };
-const char short_opts[] = "zoprcabdngheifs:t:";
+const char short_opts[] = "zoprcabdngs:heift:";
 
-
-bool wipe_source (const int fd, w_info *info) {
-
-    return true;
+void usage() {
+    fprintf(stdout, "Usage: %s [FILES] [FLAGS]", PROG_NAME);
+    fprintf(stdout, "\
+\n\
+Wipe algorithm flags:\n\
+    -z                  overwrite bytes from wile with only zeros\n\
+    -o                  overwrite bytes from wile with only ones\n\
+    -p                  overwrite bytes from wile with zeros and ones (randomly)\n\
+    -r                  use 'Russian GOST R 50739-95' method to wipe the file\n\
+    -c                  use 'US Air Force SSI-5020' method to wipe the file\n\
+    -a                  use 'US Army 380-19' method to wipe the file\n\
+    -b                  use 'British HMG IS5' method to wipe the file\n\
+    -d                  use 'US DoD 5220.22-M (ECE)' method to wipe the file\n\
+    -n                  use 'Gutmann' method to wipe the file\n\
+    -g                  use 'Pfitzner' method to wipe the file\n\
+    -s --source=SPATH   overwrite bytes from file with bytes from SPATH file\n\
+\n\
+Other flags:\n\
+    -h --help           display this menu\n\
+    -e                  delete file after it's been succesfully wiped\n\
+    -i                  get enriched information (like status bar)\n\
+    -f                  try and force open a file\n\
+    -t --first=N        wipe first and last N bytes of file\n\
+\n\
+Only one wipe agorithm flag can be set. If more than one are set, the last one set will be selected.\n\
+If no wipe algorithm flag is set, -z flag will be defaulty set.");
+    exit(EXIT_SUCCESS);
 }
+
 bool wipe_zeros (const int fd, w_info *info) {
 
     return true;
@@ -133,65 +156,46 @@ bool wipe_gutmann (const int fd, w_info *info) {
     return true;
 }
 
-void usage() {
-    fprintf(stdout, "Usage: %s [FILES] [FLAGS]", PROG_NAME);
-    fprintf(stdout, "\
-\n\
-Wipe algorithm flags:\n\
-    -z                  overwrite bytes from wile with only zeros\n\
-    -o                  overwrite bytes from wile with only ones\n\
-    -p                  overwrite bytes from wile with zeros and ones (randomly)\n\
-    -r                  use 'Russian GOST R 50739-95' method to wipe the file\n\
-    -c                  use 'US Air Force SSI-5020' method to wipe the file\n\
-    -a                  use 'US Army 380-19' method to wipe the file\n\
-    -b                  use 'British HMG IS5' method to wipe the file\n\
-    -d                  use 'US DoD 5220.22-M (ECE)' method to wipe the file\n\
-    -n                  use 'Gutmann' method to wipe the file\n\
-    -g                  use 'Pfitzner' method to wipe the file\n\
-    -s --source=SPATH   overwrite bytes from file with bytes from SPATH file\n\
-\n\
-Other flags:\n\
-    -h --help           display this menu\n\
-    -t --first=N        wipe first and last N bytes of file\n\
-\n\
-Only one wipe agorithm flag can be set. If more than one are set, the last one set will be selected.\n\
-If no wipe algorithm flag is set, -z flag will be defaulty set.");
-    exit(EXIT_SUCCESS);
-}
 
 bool failed_here(w_info *info, const char *err_msg) {
-    info->end_time = time(NULL);
-    
-    info->str_errors_buf = realloc(info->str_errors_buf, sizeof(char *) * ++(info->buf_length));
-    info->str_errors_buf[info->buf_length - 1] = malloc(strlen(err_msg) + 1);
-    strcpy(info->str_errors_buf[info->buf_length - 1], err_msg);
-
+    info->end_tm = time(NULL);
     info->wipe_failed = true;
+    
+    info->errors = realloc(info->errors, sizeof(char *) * ++(info->n_errors));
+    info->errors[info->n_errors - 1] = malloc(strlen(err_msg) + 1);
+    strcpy(info->errors[info->n_errors - 1], err_msg);
+
     return false;
 }
 
+
 bool wipe_file(const char *path, w_info *info) {
-    info->prec = 0;
-    info->buf_length = 0;
-    info->str_errors_buf = NULL;
-    info->start_tm = time(NULL);
+    info[0] = (w_info) {
+        false,      //wipe_failed
+        NULL,       //errors
+        0,          //n_errors
+        0,          //completed_prec
+        time(NULL), //start_tm
+        time(NULL)  //end_tm
+    };
+
     fprintf("Wiping file: '%s'", path);
 
+    // if the file can't be opened, try and change the users permissions and try opening it again
     int fd;
     if ((fd = open(path, O_RDWR)) == -1
         && errno == EACCES && options.force_open
         && (chmod(path, S_IRUSR | S_IWUSR) == -1 || (fd = open(path, O_RDWR)) == -1)
-        || fd == -1) return failed_here(info, "Caon't open file or change it's mode!");
+        || fd == -1) return failed_here(info, "Couldn't open the file!");
 
-    struct stat sb;;
-    if (fstat(fd, &sb) == -1) return failed_here(info, "Can't get file size!");
-    options.file_size_bytes = sb.st_blocks * 512;
+    struct stat sb;
+    if (fstat(fd, &sb) == -1) return failed_here(info, "Couldn't get file size!");
 
+    // socket or fifo files can't be overwriten
     if (S_ISSOCK(sb.st_mode) || S_ISFIFO(sb.st_mode)) return failed_here(info, "File is either a Socket or a FIFO (named pipe).");
 
 
-    return !(  wipe_with == USE_SOURCE      && !wipe_source(fd, info)
-             ||wipe_with == USE_ZEROS       && !wipe_zeros(fd, info)
+    return !(  wipe_with == USE_ZEROS       && !wipe_zeros(fd, info)
              ||wipe_with == USE_ONES        && !wipe_ones(fd, info)
              ||wipe_with == USE_PSEUDO      && !wipe_pseudo(fd, info)
              ||wipe_with == USE_GOST        && !wipe_gost(fd, info)
@@ -206,36 +210,43 @@ bool wipe_file(const char *path, w_info *info) {
 void display_wipe_info (char * const argv[], const w_info *fs_info, const int n_fs_info) {
 
     for(int i = 0; i < n_fs_info; i++) {
-            fprintf(stdout, "File: '%s'\n", argv[i + options.firs_fp]);
+            fprintf(stdout, "%d)\nFile: '%s'\n", i + 1, argv[i + first_fp]);
 
-            if (fs_info[i].wipe_failed) fprintf(stdout, "File couldn't be wiped: %s\n", fs_info[i].str_errors_buf);
+            fprintf(stdout, "%s\n", (fs_info[i].wipe_failed)? "Wipe failed!\n" : "Succesfull wipe!\n");
+
+            if (fs_info[i].wipe_failed) {
+                fprintf(stdout, "Failure reasons: \n");
+                for (int j = 0; j < fs_info[i].n_errors; j++) {
+                    fprintf(stdout, "- %s\n", fs_info[i].errors[j]);
+                }
+            }
             
-            fprintf(stdout, "Started at: %d\n", fs_info[i].start_tm);
+            fprintf(stdout, "Completed: %d\%\n\n",  fs_info[i].completed_prec);
             
-            fprintf(stdout, "%s at: %d\n", (fs_info[i].wipe_failed)? "Failed" : "Finished", fs_info[i].end_time);
+            fprintf(stdout, "Started at: %d\n", ctime(fs_info[i].start_tm));
             
-            fprintf(stdout, "Completed: %d\%\n\n",  fs_info[i].prec);
+            fprintf(stdout, "Ended at: %d\n", ctime(fs_info[i].end_tm));
     }
 }
 
 void free_info (w_info **fs_info, const int n) {
     for (int i = 0; i < n; i++) {
-        for (int j = 0; j < (*fs_info)[i].buf_length; j++) free((*fs_info)->str_errors_buf[j]);
-        free((*fs_info)[i].str_errors_buf);
+        for (int j = 0; j < (*fs_info)[i].n_errors; j++) free((*fs_info)->errors[j]);
+        free((*fs_info)[i].errors);
     }
     free((*fs_info));
 }
 
 int wipe_files (const int argc, char *argv[]) {
 
-    options.last_fp = options.firs_fp;
-    while(argv[++options.last_fp] != NULL);
+    last_fp = first_fp;
+    while(argv[++last_fp] != NULL);
 
     bool one_failed = false;
-    int n_fs_info = options.last_fp - options.firs_fp;
+    int n_fs_info = last_fp - first_fp;
     w_info *fs_info = malloc((n_fs_info) * sizeof(w_info));
-    for (int i = options.firs_fp; i < options.last_fp; i++) {
-        one_failed = (wipe_file(argv[i], &fs_info[i - options.firs_fp]))? one_failed : true;
+    for (int i = first_fp; i < last_fp; i++) {
+        one_failed = (wipe_file(argv[i], &fs_info[i - first_fp]))? one_failed : true;
     }
 
     display_wipe_info(argv, fs_info, n_fs_info);
@@ -246,10 +257,10 @@ int wipe_files (const int argc, char *argv[]) {
         char c;
         fscanf(stdin, "%c", &c);
         if (tolower(c) == 'y') {
-            int last_failed_fp = options.firs_fp;
-            for (int i = 0; i < options.last_fp - options.firs_fp; i++) 
+            int last_failed_fp = first_fp;
+            for (int i = 0; i < last_fp - first_fp; i++) 
                 if (fs_info[i].wipe_failed) 
-                    argv[last_failed_fp++] = argv[i + options.firs_fp];
+                    argv[last_failed_fp++] = argv[i + first_fp];
 
             argv[last_failed_fp] = NULL;
 
@@ -266,80 +277,50 @@ int wipe_files (const int argc, char *argv[]) {
 
 }
 
-void set_flags (const int argc, char * const argv[]) {
+void set_options (const int argc, char * const argv[]) {
     wipe_with = USE_NONE;
+    options = (Options) {
+        false,  // delete_after
+        false,  // displ_info
+        false,  // force_open
+        0,      // first
+    };
 
     int opt;
     while((opt = getopt_long(argc, argv, short_opts, long_opts, NULL)) != -1) {
         switch (opt) {
-            case 'z':
-                wipe_with = USE_ZEROS;
-                break;
-            case 'o':
-                wipe_with = USE_ONES;
-                break;
-            case 'p':
-                wipe_with = USE_PSEUDO;
-                break;
-            case 'r':
-                wipe_with = USE_GOST;
-                break;
-            case 'c':
-                wipe_with = USE_AIRFORCE;
-                break;
-            case 'a':
-                wipe_with = USE_ARMY;
-                break;
-            case 'b':
-                wipe_with = USE_HMG;
-                break;
-            case 'd':
-                wipe_with = USE_DOD;
-                break;
-            case 'n':
-                wipe_with = USE_PFITZNER;
-                break;
-            case 'g':
-                wipe_with = USE_GUTMANN;
-                break;
-            case 's':
-                wipe_with = USE_SOURCE;
-                FAIL_IF((options.fd_src = open(optarg + 1, O_RDONLY)) == -1, "Can't open source file!");
-                fprintf(stdout, "Overwriting by coping bytes from source file: \'%s\'\n(No wipe algorithm will be used)\n", optarg + 1);
-                break;
-
-            //others
-            case 'h':
-                usage();
-                break;
-            case 'e':
-                options.delete_after = true;
-                break;
-            case 'i':
-                options.displ_info = true;
-                break;
-            case 'f':
-                options.force_open = true;
-                break;
-            case 't':
-                options.first_last = atoi(optarg + 1);
-                FAIL_IF(options.first_last == 0 || options.first_last != strtol(optarg + 1, NULL, 10), "Argument to -t/--first must be a positive intiger!");
-                break;
+            // wipe methods
+            case 'z':   wipe_with = USE_ZEROS;          break;
+            case 'o':   wipe_with = USE_ONES;           break;
+            case 'p':   wipe_with = USE_PSEUDO;         break;
+            case 'r':   wipe_with = USE_GOST;           break;
+            case 'c':   wipe_with = USE_AIRFORCE;       break;
+            case 'a':   wipe_with = USE_ARMY;           break;
+            case 'b':   wipe_with = USE_HMG;            break;
+            case 'd':   wipe_with = USE_DOD;            break;
+            case 'n':   wipe_with = USE_PFITZNER;       break;
+            case 'g':   wipe_with = USE_GUTMANN;        break;
+            // other flags
+            case 'h':   usage();                        break;
+            case 'e':   options.delete_after = true;    break;
+            case 'i':   options.displ_info = true;      break;
+            case 'f':   options.force_open = true;      break;
+            case 't':   FAIL_IF((options.first = atoi(optarg + 1)) <= 0 
+                                || options.first != strtol(optarg + 1, NULL, 10)
+                                , "Argument to -t/--first must be a positive intiger!"); break;
         }
     }
 
     if (wipe_with == USE_NONE) wipe_with = USE_ZEROS;
 
-    FAIL_IF(optind == argc, "Missing file opperand!");
-    options.firs_fp = optind;
+    FAIL_IF((first_fp = optind) == argc, "Missing file opperand!");
 }
 
 int main (int argc, char **argv) {
     time_t start_tm = time(NULL);
 
-    set_flags(argc, argv);
+    set_options(argc, argv);
 
-    
     int fin_status;
     while((fin_status = wipe_files(argc, argv)) == FALIURE_RETRY);
     
@@ -349,7 +330,6 @@ int main (int argc, char **argv) {
     struct tm *lt = localtime(&diff_tm);
     fprintf(stdout, "Total time elapsed: %d days|%d hrs|%d mins|%s secs\n", lt->tm_mday, lt->tm_hour, lt->tm_min, lt->tm_sec);
 
-    FAIL_IF(options.fd_src != -1 && close(options.fd_src) == -1, "Can't close src file!");
 
     return 0;
 }
